+++ crates/rust-client/src/lib.rs
@@ -39,6 +39,7 @@ pub mod assets {
 /// Provides authentication-related types and functionalities for the Miden
 /// rollup network.
 pub mod auth {
+    pub use miden_lib::AuthScheme;
     pub use miden_objects::accounts::AuthSecretKey;
     pub use miden_tx::auth::{BasicAuthenticator, TransactionAuthenticator};
 }
@@ -54,6 +55,7 @@ pub mod blocks {
 pub mod crypto {
     pub use miden_objects::{
         crypto::{
+            dsa::rpo_falcon512::SecretKey,
             merkle::{
                 InOrderIndex, LeafIndex, MerklePath, MmrDelta, MmrPeaks, MmrProof, SmtLeaf,
                 SmtProof,
@@ -119,6 +121,7 @@ pub struct Client<R: FeltRng> {
     tx_executor: TransactionExecutor,
 }
 
+/// Construction and access methods.
 impl<R: FeltRng> Client<R> {
     // CONSTRUCTOR
     // --------------------------------------------------------------------------------------------
@@ -142,7 +145,7 @@ impl<R: FeltRng> Client<R> {
     ///
     /// # Errors
     ///
-    /// Returns an error if the client could not be instantiated.
+    /// Returns an error if the client couldn't be instantiated.
     pub fn new(
         rpc_api: Box<dyn NodeRpcClient + Send>,
         rng: R,